install scripts wip; refactoring
